Add Mikado pleasant logo
Mikado pleasant
This is the endemic bird species of Taiwan-the Mikado pheasant. The Mikado pheasants live in high mountains with an altitude of 1,800 to 3,800 meters, which are often surrounded by clouds and fog. They are shy and alert. They can only detect their gorgeous appearance and graceful posture when the fog appears or the edge of the forest after light rain. As "the king in the mist".